+++ src/main/java/com/example/yammiebackend/Repository/OrderNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.yammiebackend.Repository;
 
 public class OrderNotFoundException extends RuntimeException {
 
-    public OrderNotFoundException(Long id) {
+    public OrderNotFoundException(int id) {
         super("Could not find order " + id);
     }
 }
