+++ src/main/java/com/example/yammiebackend/Repository/OrderRepository.java
@@ -3,5 +3,6 @@ package com.example.yammiebackend.Repository;
 import com.example.yammiebackend.Model.PlacedOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<PlacedOrder, Long> {
+public interface OrderRepository extends JpaRepository<PlacedOrder, Integer> {
+
 }
